Started Computer Vision Project
Tested different edge detection algorithms.
Something fucky is going on with the drawing lines thing I have now.

A few notes:
Sobel [++ ]
Canny {+++] Good at finding edges, but not necessarily our corners.

It might be wise to think of the algorithm as trying to find negative
space and not points on a corner.